refactor: create custom hook for handling calculation
