+++ src/components/ShareCalculator/SellReport.tsx
@@ -1,25 +1,35 @@
-interface SellSummaryProps{
-  totalAmount : number,
-  commission : number,
-  sebonFee : number,
-  dpCharge : number,
-  totalAmountReceiveable: number,
-  capitalGainTax : number,
-  capitalGainTaxPercentage :number,
-  profitOrLoss : number,
-  nepseCommission : number,
-  sebonRegularityFee : number,
+interface SellReportProps {
+  totalAmount: number;
+  commission: number;
+  sebonFee: number;
+  dpCharge: number;
+  totalAmountReceiveable: number;
+  capitalGainTax: number;
+  capitalGainTaxPercentage: number;
+  profitOrLoss: number;
+  nepseCommission: number;
+  sebonRegularityFee: number;
 }
 
-
-const SellSummary = ({totalAmount, commission, sebonFee, dpCharge, totalAmountReceiveable, capitalGainTax, capitalGainTaxPercentage, profitOrLoss, nepseCommission, sebonRegularityFee} : SellSummaryProps) => {
-    return (
-        <>
-        <div className="p-8 border border-gray-300 dark:border-gray-600 rounded-lg">
+const SellReport = ({
+  totalAmount,
+  commission,
+  sebonFee,
+  dpCharge,
+  totalAmountReceiveable,
+  capitalGainTax,
+  capitalGainTaxPercentage,
+  profitOrLoss,
+  nepseCommission,
+  sebonRegularityFee,
+}: SellReportProps) => {
+  return (
+    <>
+      <div className="p-8 border border-gray-300 dark:border-gray-600 rounded-lg">
         <h3 className="text-lg font-semibold mb-4">Financial Summary</h3>
         <table className="w-full border-collapse border border-gray-300 dark:border-gray-600 rounded-md">
           <tbody>
-           <tr className="border-b">
+            <tr className="border-b">
               <td className="py-2 px-4 font-semibold">Total Amount</td>
               <td className="py-2 px-4">{totalAmount}</td>
             </tr>
@@ -48,22 +58,20 @@ const SellSummary = ({totalAmount, commission, sebonFee, dpCharge, totalAmountRe
               <td className="py-2 px-4">{capitalGainTax}</td>
             </tr>
             <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">
-                Profit / Loss (Rs)
-              </td>
+              <td className="py-2 px-4 font-semibold">Profit / Loss (Rs)</td>
               <td className="py-2 px-4">{profitOrLoss}</td>
             </tr>
             <tr>
               <td className="py-2 px-4 font-semibold">
-                Commission Amount includes NEPSE Commission Rs {nepseCommission} & SEBON
-                Regularity Fee Rs {sebonRegularityFee}
+                Commission Amount includes NEPSE Commission Rs {nepseCommission}{" "}
+                & SEBON Regularity Fee Rs {sebonRegularityFee}
               </td>
             </tr>
           </tbody>
         </table>
       </div>
-        </>
-    );
-}
+    </>
+  );
+};
 
-export default SellSummary;
+export default SellReport;
