+++ src/components/ShareCalculator/index.tsx
@@ -1,79 +1,40 @@
 /**
- * External dependencies.
+ * Internal dependencies.
  */
-import React, { useState } from "react";
+import SellReport from "./SellReport";
+import BuyReport from "./BuyReport";
+import useShareCalculator from "./useShareCalculator";
 
 const ShareCalculator = () => {
-  const [transactionType, setTransactionType] = useState<string>("buy");
-  const [shareQuantity, setShareQuantity] = useState<number>(0);
-  const [purchasePrice, setPurchasePrice] = useState<number>(0);
-  const [isWACC, setIsWACC] = useState<boolean>(false);
-  const [investorType, setInvestorType] = useState<string>("individual");
-  const [taxRate, setTaxRate] = useState<number>(5);
-
-  // Reset to default value.
-  const handleTransactionTypeChange = (
-    e: React.ChangeEvent<HTMLSelectElement>
-  ) => {
-    const _transactionType = e.target.value ?? "";
-    setTransactionType(_transactionType);
-    setShareQuantity(0);
-    setPurchasePrice(0);
-    setIsWACC(false);
-  };
-
-  const handleShareQuantityChange = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const _shareQuantity = Number(
-      (e.target.value ?? "").replace(/[^0-9]/g, "")
-    );
-    setShareQuantity(_shareQuantity);
-  };
-
-  const handlePurchasePriceChange = (
-    e: React.ChangeEvent<HTMLInputElement>
-  ) => {
-    const _purchasePrice = Number(
-      (e.target.value ?? "").replace(/[^0-9.]/g, "")
-    );
-    setPurchasePrice(_purchasePrice);
-  };
-
-  const handleIsWaccChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const _isWacc = e.target.checked ?? false;
-    setIsWACC(_isWacc);
-  };
-
-  const handleInvestorTypeChange = (
-    e: React.ChangeEvent<HTMLSelectElement>
-  ) => {
-    const _investorType = e.target.value ?? "individual";
-    const _filteredInvestorType =
-      _investorType === "individual" || _investorType === "institutional"
-        ? _investorType
-        : "individual";
-
-    setInvestorType(_filteredInvestorType);
-  };
-
-  const handleTaxRateChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
-    const _taxRate = Number((e.target.value ?? 7.5).replace(/[^0-9.]/g, ""));
-    setTaxRate(_taxRate);
-  };
-
-  const handleCalculate = () => {
-    console.log("Calculating...");
-  };
-
-  const handleClear = () => {
-    setTransactionType("buy");
-    setShareQuantity(0);
-    setPurchasePrice(0);
-    setIsWACC(false);
-    setInvestorType("individual");
-    setTaxRate(5);
-  };
+  const {
+    transactionType,
+    shareQuantity,
+    purchasePrice,
+    sellingPrice,
+    investorType,
+    taxRate,
+    isWACC,
+    totalAmount,
+    commission,
+    sebonFee,
+    dpCharge,
+    capitalGainTax,
+    totalAmountPayable,
+    capitalGainTaxPercentage,
+    costPerShare,
+    profitOrLoss,
+    nepseCommission,
+    sebonRegularityFee,
+    totalAmountReceiveable,
+    handleClear,
+    handleIsWaccChange,
+    handleTaxRateChange,
+    handleTransactionTypeChange,
+    handleShareQuantityChange,
+    handlePurchasePriceChange,
+    handleSellingPriceChange,
+    handleInvestorTypeChange,
+  } = useShareCalculator();
 
   return (
     <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
@@ -109,6 +70,16 @@ const ShareCalculator = () => {
                 type="text"
                 value={purchasePrice}
                 onChange={handlePurchasePriceChange}
+                className="border border-gray-300 dark:border-gray-600 rounded-md px-3 py-2 focus:outline-none focus:ring focus:border-blue-300 w-full bg-white dark:bg-gray-900 text-gray-900 dark:text-white"
+                placeholder="Enter purchase price"
+              />
+              <label className="block font-bold mb-1">
+                Selling Price (Rs):
+              </label>
+              <input
+                type="text"
+                value={sellingPrice}
+                onChange={handleSellingPriceChange}
                 className="border border-gray-300 dark:border-gray-600 rounded-md px-3 py-2 focus:outline-none focus:ring focus:border-blue-300 w-full bg-white dark:bg-gray-900 text-gray-900 dark:text-white"
                 placeholder="Enter purchase price"
               />
@@ -182,12 +153,6 @@ const ShareCalculator = () => {
 
         <div className="flex justify-end space-x-4 mt-5">
           <button
-            onClick={handleCalculate}
-            className="bg-blue-500 text-white px-4 py-2 rounded-md hover:bg-blue-600 focus:outline-none focus:ring focus:border-blue-300"
-          >
-            Calculate
-          </button>
-          <button
             onClick={handleClear}
             className="px-4 py-2 rounded-md bg-gray-500 hover:bg-gray-400 focus:outline-none focus:ring focus:border-blue-300 text-white"
           >
@@ -195,48 +160,33 @@ const ShareCalculator = () => {
           </button>
         </div>
       </div>
-      <div className="p-8 border border-gray-300 dark:border-gray-600 rounded-lg">
-        <h3 className="text-lg font-semibold mb-4">Financial Summary</h3>
-        <table className="w-full border-collapse border border-gray-300 dark:border-gray-600 rounded-md">
-          <tbody>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">Total Amount</td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">Commission</td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">SEBON FEE</td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">DP Charge</td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">
-                Total Amount Payable (Rs)
-              </td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr className="border-b">
-              <td className="py-2 px-4 font-semibold">
-                Cost Price Per Share (Rs)
-              </td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-            <tr>
-              <td className="py-2 px-4 font-semibold">
-                Commission Amount includes NEPSE Commission Rs - & SEBON
-                Regularity Fee Rs -
-              </td>
-              <td className="py-2 px-4">-</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
+
+      {transactionType === "buy" && (
+        <BuyReport
+          totalAmount={totalAmount}
+          commission={commission}
+          sebonFee={sebonFee}
+          dpCharge={dpCharge}
+          totalAmountPayable={totalAmountPayable}
+          costPerShare={costPerShare}
+          nepseCommission={nepseCommission}
+          sebonRegularityFee={sebonRegularityFee}
+        />
+      )}
+      {transactionType === "sell" && (
+        <SellReport
+          totalAmount={totalAmount}
+          commission={commission}
+          sebonFee={sebonFee}
+          dpCharge={dpCharge}
+          totalAmountReceiveable={totalAmountReceiveable}
+          capitalGainTax={capitalGainTax}
+          capitalGainTaxPercentage={capitalGainTaxPercentage}
+          profitOrLoss={profitOrLoss}
+          nepseCommission={nepseCommission}
+          sebonRegularityFee={sebonRegularityFee}
+        />
+      )}
     </div>
   );
 };
